requirements.txt
streamlit==1.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
plotly==5.18.0
psycopg2-binary==2.9.9
pyjpboatrace==0.2.0
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
